requirements.txt
streamlit==1.45.1
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3
matplotlib-inline==0.1.7
tensorflow==2.16.1
Pillow==10.3.0
joblib==1.4.2